benchmark code cleanup
