--- gflog-benchmark/src/main/java/com/epam/deltix/gflog/benchmark/LatencyBenchmark.java
@@ -1,146 +1,0 @@
-package com.epam.deltix.gflog.benchmark;
-
-import com.epam.deltix.gflog.api.Log;
-import com.epam.deltix.gflog.api.LogFactory;
-import com.epam.deltix.gflog.benchmark.util.BenchmarkDescriptor;
-import com.epam.deltix.gflog.benchmark.util.BenchmarkState;
-import com.epam.deltix.gflog.benchmark.util.BenchmarkUtil;
-
-import java.util.ArrayList;
-import java.util.LinkedHashMap;
-import java.util.List;
-import java.util.Map;
-
-import static com.epam.deltix.gflog.benchmark.util.BenchmarkUtil.*;
-
-/**
- * Shows safepoints: -XX:+UnlockDiagnosticVMOptions -XX:+PrintGCDateStamps -XX:+PrintGCApplicationStoppedTime
- * Shows details: -XX:+PrintSafepointStatistics -XX:PrintSafepointStatisticsCount=1
- * Adjusts interval: -XX:GuaranteedSafepointInterval=1200000
- */
-public class LatencyBenchmark {
-
-    // 100 bytes + (line separator 1-2)
-    // 2020-10-01 14:25:58.310 INFO '123456789012345678901234567890' [12345678901234567890123] Hello world!
-
-    private static final String[] CONFIGS = {
-            "noop",
-            /*"console-direct",
-            "console-wrapper",*/
-            "file",
-            "rolling-file"
-    };
-
-    public static void main(final String[] args) throws Exception {
-        final Map<String, BenchmarkDescriptor> all = benchmarks();
-        final LatencyBenchmarkRunner runner = new LatencyBenchmarkRunner(all);
-
-        runner.run(args);
-    }
-
-    private static void entry1Arg(final BenchmarkState state) {
-        Holder.LOG.info().append(MESSAGE).commit();
-    }
-
-    private static void template1Arg(final BenchmarkState state) {
-        Holder.LOG.info(MESSAGE);
-    }
-
-    public static void entry5Args(final BenchmarkState state) {
-        Holder.LOG.info()
-                .append("Some array: [")
-                .append(state.arg1).append(',')
-                .append(state.arg2).append(',')
-                .append(state.arg3).append(',')
-                .append(state.arg4).append(',')
-                .append(state.arg5).append(']')
-                .commit();
-    }
-
-    public static void template5Args(final BenchmarkState state) {
-        Holder.LOG.info("Some array: [%s,%s,%s,%s,%s]")
-                .with(state.arg1)
-                .with(state.arg2)
-                .with(state.arg3)
-                .with(state.arg4)
-                .with(state.arg5);
-    }
-
-    public static void entry10Args(final BenchmarkState state) {
-        Holder.LOG.info()
-                .append("Some array: [")
-                .append(state.arg1).append(',')
-                .append(state.arg2).append(',')
-                .append(state.arg3).append(',')
-                .append(state.arg4).append(',')
-                .append(state.arg5).append(',')
-                .append(state.arg6).append(',')
-                .append(state.arg7).append(',')
-                .append(state.arg8).append(',')
-                .append(state.arg9).append(',')
-                .append(state.arg10).append(']')
-                .commit();
-    }
-
-    public static void template10Args(final BenchmarkState state) {
-        Holder.LOG.info("Some array: [%s,%s,%s,%s,%s,%s,%s,%s,%s,%s]")
-                .with(state.arg1)
-                .with(state.arg2)
-                .with(state.arg3)
-                .with(state.arg4)
-                .with(state.arg5)
-                .with(state.arg6)
-                .with(state.arg7)
-                .with(state.arg8)
-                .with(state.arg9)
-                .with(state.arg10);
-    }
-
-    private static void entryException(final BenchmarkState state) {
-        Holder.LOG.info().append("Exception: ").append(state.exception).commit();
-    }
-
-    private static void templateException(final BenchmarkState state) {
-        Holder.LOG.info("Exception: %s").with(state.exception);
-    }
-
-    private static Map<String, BenchmarkDescriptor> benchmarks() {
-        final List<BenchmarkDescriptor> benchmarks = new ArrayList<>();
-        final Runnable noop = () -> {
-        };
-
-        benchmarks.add(new BenchmarkDescriptor("baseline", noop, noop, LatencyBenchmarkRunner::baseline));
-        benchmarks.add(new BenchmarkDescriptor("timestamp", noop, noop, LatencyBenchmarkRunner::timestamp));
-
-        for (final String config : CONFIGS) {
-            final Runnable prepare = () -> BenchmarkUtil.prepare(config, ENCODING);
-            final Runnable cleanup = BenchmarkUtil::cleanup;
-
-            benchmarks.add(new BenchmarkDescriptor("entry1Arg-" + config, prepare, cleanup, LatencyBenchmark::entry1Arg));
-            benchmarks.add(new BenchmarkDescriptor("entry5Args-" + config, prepare, cleanup, LatencyBenchmark::entry5Args));
-            benchmarks.add(new BenchmarkDescriptor("entry10Args-" + config, prepare, cleanup, LatencyBenchmark::entry10Args));
-            benchmarks.add(new BenchmarkDescriptor("entryException-" + config, prepare, cleanup, LatencyBenchmark::entryException));
-
-            benchmarks.add(new BenchmarkDescriptor("template1Arg-" + config, prepare, cleanup, LatencyBenchmark::template1Arg));
-            benchmarks.add(new BenchmarkDescriptor("template5Args-" + config, prepare, cleanup, LatencyBenchmark::template5Args));
-            benchmarks.add(new BenchmarkDescriptor("template10Args-" + config, prepare, cleanup, LatencyBenchmark::template10Args));
-            benchmarks.add(new BenchmarkDescriptor("templateException-" + config, prepare, cleanup, LatencyBenchmark::templateException));
-        }
-
-        final Map<String, BenchmarkDescriptor> map = new LinkedHashMap<>();
-
-        for (final BenchmarkDescriptor benchmark : benchmarks) {
-            final String name = benchmark.getName();
-            map.put(name, benchmark);
-        }
-
-        return map;
-    }
-
-    private static final class Holder {
-
-        private static final Log LOG = LogFactory.getLog(LOGGER);
-
-    }
-
-}
