+++ gflog-benchmark/src/main/java/com/epam/deltix/gflog/benchmark/gflog/GflogStubBenchmark.java
@@ -1,4 +1,4 @@
-package com.epam.deltix.gflog.benchmark;
+package com.epam.deltix.gflog.benchmark.gflog;
 
 import com.epam.deltix.gflog.api.Log;
 import com.epam.deltix.gflog.api.LogFactory;
@@ -16,9 +16,9 @@ import java.util.concurrent.TimeUnit;
 @State(Scope.Benchmark)
 @BenchmarkMode(Mode.AverageTime)
 @OutputTimeUnit(TimeUnit.NANOSECONDS)
-public class StubOverheadBenchmark {
+public class GflogStubBenchmark {
 
-    private static final Log LOG = LogFactory.getLog(StubOverheadBenchmark.class);
+    private static final Log LOG = LogFactory.getLog(GflogStubBenchmark.class);
 
 
     @Benchmark
@@ -121,7 +121,7 @@ public class StubOverheadBenchmark {
 
     public static void main(final String[] args) throws Exception {
         final Options opt = new OptionsBuilder()
-                .include(StubOverheadBenchmark.class.getName())
+                .include(GflogStubBenchmark.class.getName())
                 .build();
 
         new Runner(opt).run();
