+++ gflog-benchmark/src/main/java/com/epam/deltix/gflog/benchmark/misc/CopyMemoryBenchmark.java
@@ -1,7 +1,6 @@
-package com.epam.deltix.gflog.benchmark.util;
+package com.epam.deltix.gflog.benchmark.misc;
 
 import com.epam.deltix.gflog.core.util.Util;
-import org.openjdk.jmh.annotations.Benchmark;
 import org.openjdk.jmh.annotations.*;
 import org.openjdk.jmh.runner.Runner;
 import org.openjdk.jmh.runner.RunnerException;
