+++ gflog-benchmark/src/main/resources/com/epam/deltix/gflog/benchmark/gflog/gflog-console-direct-benchmark.xml
