+++ gflog-benchmark/src/main/resources/com/epam/deltix/gflog/benchmark/gflog/gflog-file-benchmark.xml
